Use a single source file for scala solutions
Put the file in the top level folder (package nesting is not needed here)
