--- src/main/scala/io/evilcorp/step01/Step01ScalaFunctional.scala
@@ -1,12 +1,0 @@
-package io.evilcorp.step01
-
-object Step01ScalaFunctional {
-  def apply(string : String) : String =
-    string.split(" ")
-      .map((s) => if (s.equals("nice")) "XXXX" else s)
-      .mkString(" ")
-
-  def main(args: Array[String]): Unit = {
-    println(apply("You are a nice person"))
-  }
-}
